+++ webapp/apps/dynamic/tests/test_views.py
@@ -15,6 +15,7 @@ from taxcalc import Policy
 
 START_YEAR = u'2016'
 
+
 class DynamicViewsTests(TestCase):
     ''' Test the views of this app. '''
 
@@ -80,7 +81,7 @@ class DynamicViewsTests(TestCase):
         link_idx = response.url[:-1].rfind('/')
         self.failUnless(response.url[:link_idx+1].endswith("behavior_results/"))
 
- 
+
     def test_elastic_post(self):
         #Monkey patch to mock out running of compute jobs
         import sys
@@ -184,5 +185,3 @@ class DynamicViewsTests(TestCase):
         response = self.client.get(response.url)
         # Make sure the failure message is in the response
         self.failUnless("Your calculation failed" in str(response))
-
-
