+++ webapp/apps/dynamic/tests/utils.py
@@ -19,44 +19,6 @@ from ...test_assets import *
 
 START_YEAR = u'2016'
 
-def do_micro_sim(client, reform):
-    '''do the proper sequence of HTTP calls to run a microsim'''
-    #Monkey patch to mock out running of compute jobs
-    import sys
-    from webapp.apps.taxbrain import views
-    webapp_views = sys.modules['webapp.apps.taxbrain.views']
-    webapp_views.dropq_compute = MockCompute()
-    from webapp.apps.dynamic import views
-    dynamic_views = sys.modules['webapp.apps.dynamic.views']
-    dynamic_views.dropq_compute = MockCompute(num_times_to_wait=1)
-
-    response = client.post('/taxbrain/', reform)
-    # Check that redirect happens
-    assert response.status_code == 302
-    idx = response.url[:-1].rfind('/')
-    assert response.url[:idx].endswith("taxbrain")
-    return response
-
-
-def do_micro_sim_from_file(client):
-    #Monkey patch to mock out running of compute jobs
-    import sys
-    from webapp.apps.taxbrain import views
-    webapp_views = sys.modules['webapp.apps.taxbrain.views']
-    webapp_views.dropq_compute = MockCompute()
-
-    tc_file = SimpleUploadedFile("test_reform.json", reform_text)
-    tc_file2 = SimpleUploadedFile("test_assumptions.json", assumptions_text)
-    data = {u'docfile': tc_file,
-            u'assumpfile': tc_file2,
-            u'has_errors': [u'False'],
-            u'start_year': unicode(START_YEAR), 'csrfmiddlewaretoken':'abc123'}
-
-    response = client.post('/taxbrain/file/', data)
-    # Check that redirect happens
-    assert response.status_code == 302
-    return response
-
 
 def do_behavioral_sim(client, microsim_response, pe_reform, start_year=START_YEAR):
     # Link to dynamic simulation
